move FiguredBass object to figuredBass.FiguredBass
