+++ music21/figuredBass/__init__.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 __all__ = [
     'checker',
     'examples',
+    'harmony',
     'notation',
     'possibility',
     'realizer',
@@ -11,10 +12,14 @@ __all__ = [
     'resolution',
     'rules',
     'segment',
+
+    'FiguredBass',
 ]
 
 from music21.figuredBass import checker
 from music21.figuredBass import examples
+from music21.figuredBass import harmony
+from music21.figuredBass.harmony import FiguredBass
 from music21.figuredBass import notation
 from music21.figuredBass import possibility
 from music21.figuredBass import realizer
